Create Hangman project in pycharm.
In this command line python project, a word is guessed from the list of words according to which lives gets decreasing for every wrong guess of a word and you lose a game when all six lives are gone. By every loss of a life, a hangman image gets initiated from a stages list of images with corresponding to it's life.