Create Number Guessing Game in python
In this famous command line python project, there are 3 ways to play a number guessing game from 1 to 100: easy, hard and very hard, according to which a player will get corresponding number to attempts to finish a game. Also in this program, a concept of global scope and local scope is also taken into consideration which you can see for a variable named "attempts".