Add files via upload
In this python command line project, there is a instagram follower competition between two random famous personalities of who has more followers among them. If user guesses it right, then he/she will get one point for it, but if a user guesses wrong, then his/her final score will be displayed with a message of if a user wants to play the game again or not and if a user answers no then the game will end.